modules/accelerometer_gyroscope/requirements.txt
mpu6050-raspberrypi==1.1
